cpu/em_cpuidle: Add support for AMD EPYC systems
Add support for AMD EPYC systems in the existing CPUIdle test. Also
add the functionality to verify the flags of the CPUIdle states C1 and
C2.

While at it, tweak the check for cpupower tool availabilty, so that we
don't fail the test if cpupower is able to run even though we don't
have the linux-tools package (in case we have built it from source).

Ensure that cpuidle is present on the system, and that a driver
exists.

And finally replace the function cmp() with a new function
check_match() which can even do partial matches which is needed for
AMD systems.

[gautham.shenoy@amd.com:
     * Add a new check_match function for partial matches.
     * Run the test only if cpuidle subsystem is present, and the cpuidle driver exists.
]
Signed-off-by: Dhananjay Ugwekar <dhananjay.ugwekar@amd.com>
Signed-off-by: Gautham R. Shenoy <gautham.shenoy@amd.com>