cpu/pmqos-cpu-latency: Add a test to validate cpu_dma_latency constraint
This test verifies that the cpu_dma_latency value is honored by the
cpuidle subsystem. It writes a value "x" (in ms) to the
cpu_dma_latency file and validates that no CState which has latency
greater than "x" is entered. It also validates that at least one CState
with latency less than or equal to "x" is entered during the idle
period.

[gautham.shenoy@amd.com : Made following optimizations
 * Check for the presence of cpuidle subsystem and cpuidle driver
 * Open the cpu_dma_latency dev-file only once in test
 * Create a single helper to get cstate attributes such as latency and time
 * Break down test() into multiple small helper functions
 * Change the byte-order from 'little' to sys.byteorder for portability
 * Validate that at least one valid state (latency <= cpu_dma_latency) is entered into
 * Add a test case to validate when cpu_dma_latency is greater than all C-state latencies
]

Signed-off-by: Dhananjay Ugwekar <dhananjay.ugwekar@amd.com>
Signed-off-by: Gautham R. Shenoy <gautham.shenoy@amd.com>