cpu/hwmon-powercap: Add a test to validate powercap set via hwmon
Add a testcase to validate that the package powercap set via the HSMP
hwmon interface is being correctly enforced.

The test loads a socket with 100% busy workload and increases the
powercap limit from a minimum of power1_cap_max/2 to power1_cap_max
(advertised by hsmp-hwmon) in steps of 25W and validates that the
socket power observed via the RAPL pkg-energy counters remains within
the specified powercap.

Signed-off-by: Gautham R. Shenoy <gautham.shenoy@amd.com>