cpu/pmqos-cpu-latency: Fixes to eliminate false positives
pmqos-cpu-latency test in its current form can cause false positives
due to the following reasons

* When the target CPU is busy during the period of observation, no
  idle states will be entered. This is flagged as an error which is
  incorrect.

* If the target CPU was already idle before setting the
  pmqos-cpu-latency, the subsequent test will fail since it may have
  already been in a deeper state than what was desired by the PMQOS.

Thus fix this by * Choosing a random target CPU instead of CPU0 each
time.  * Retry the test when the CPU was busy throughout the
observation duration.  * Ensure that the target CPU is woken up soon
after setting the pmqos-cpu-latency, but before we perform the test.

Signed-off-by: Gautham R. Shenoy <gautham.shenoy@amd.com>
Reviewed-by: Swapnil Sapkal <swapnil.sapkal@amd.com>