Run cpufreq kselftest only if driver is acpi-cpufreq
On AMD systems, run cpufreq kselftest only if the driver is acpi-cpufreq
as amd-pstate driver active mode does not support ondemand and userspace
governors and cpufreq kselftest try to write scaling_governor with these
values which causes the testcase t fail. Skip this testcase if the
driver is not acpi-cpufreq.

Tested-by: Amandeep Kaur Longia <AmandeepKaur.Longia@amd.com>
Signed-off-by: Swapnil Sapkal <swapnil.sapkal@amd.com>